query_server/requirements.txt
flask
pycurl
pyaml